novel-agent-day1/requirements.txt
openai>=1.0
python-dotenv>=1.0.0